+++ CHANGELOG.md
@@ -6,6 +6,13 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic
 Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [Unreleased]
+
+### Fixed
+
+- Fixed getting the local time offset on Linux. The `time` crate normally does
+  not allow this.
+
 ## [0.3.0] - 2023-03-21
 
 ### Added